+++ Helper/FlipFilters/ProfitPercentageDetailedFlipFilter.cs
@@ -7,7 +7,7 @@ namespace Coflnet.Sky.Commands.Shared
 {
     public class ProfitPercentageDetailedFlipFilter : NumberDetailedFlipFilter
     {
-        protected override Expression<Func<FlipInstance, long>> GetSelector()
+        protected override Expression<Func<FlipInstance, double>> GetSelector()
         {
             return (f) => (long)f.ProfitPercentage;
         }
